Amend `cilk_reducer` token to preserve correct type highlighting
The `cilk_reducer` keyword is now recognized by CC Mode as a "declaration
hang-on with nontype parameters in parentheses" instead of a "type modifier".
Consider a line:

    int cilk_reducer(zero, plus) sum = 0;

The change allows CC Mode to correctly identify `int` as a type.  Before this
commit, the command in the monoid function list would trigger de-fontification
of `int`.